refactor: Moving "SparkDashboard" to a seperate file
